api/requirements.txt
Django
djangorestframework
markdown
django-filter
flake8
psycopg2
requests